Add files via upload
There are 3 main scripts that I used in this code. The first is the shinyapp code that is the user interface that uses the model and base data that I used in the other two codes to then create the Shiny app to allow the user to manipulate data. The other file is the model file that I used to get the raw data and create an H2O model that ran for 8hrs (I used 25400 seconds and not the 7200 that is in the script). Then the Create_App_Data coerces the raw data set and gets the data into a way that makes it easy to read in the Shiny App.